Indexer/main/requirements.txt
click==8.1.3
joblib==1.2.0
kafka-python==2.0.2
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
six==1.16.0
tqdm==4.65.0
